Move some of the validation / normalization function into util module
Also move some unit tests around
